requirements.txt
streamlit
torch
transformers
sentence-transformers
numpy
nltk
pymupdf
scikit-learn
python-docx